Version2

+++ app.py
@@ -10,8 +10,10 @@ import speech_recognition as sr
 import threading
 import send_sms
 from twilio.rest import Client
+from collections import deque
 
 def listen(r,m,audio):
+    global numberTextToAnalyze
     while listenTrue:
         sem.acquire()
         with m as source:
@@ -28,53 +30,67 @@ def listen(r,m,audio):
         sem2.release()
 
 def analyzer():
+    global numberTextToAnalyze
     print("Starting analysis")
     while listenTrue:
-        start_time = time.time()
-        time.sleep(30.0) #sleep for at least 10 seconds between each analysis
-        sem2.acquire()
-        f = open("output.txt", "r+")
-        text = f.read()
-        f.seek(0)
-        f.truncate()
-        f.close()
-        sem2.release()
-        print("Analyzing...", datetime.datetime.now(),'\n',text)
-        text = text.lower()
-        for word in stopWords:
-            if word in text.lower():
-                print("Keyword Found")
-                time.sleep(25.0)
-                sem2.acquire()
-                f = open("output.txt", "r+")
-                text += f.read()
-                f.seek(0)
-                f.truncate()
-                f.close()
-                sem2.release()
-                send_sms.send(text) #Sends recorded message to list of numbers
-                text = ""
-                print("How long it took to form the message from hearing the keyword: ", time.time()-start_time)
+            time.sleep(30)
+            sem2.acquire()
+            f = open("output.txt", "r+")
+            text = f.read()
+            f.seek(0)
+            f.truncate()
+            f.close()
+            sem2.release()
+            print('\n',"Analyzing...", datetime.datetime.now(),'\n',text)
+            text = text.lower()
+            last = ""
+            if text != "":
+                last = text
+            now = datetime.datetime.now()
+            if now.minute == 1:
+                send_sms.sendToMaster(last)
+                time.sleep(60)
+            if now.minute <= 30 and now.minute >= 5:
+                for number in numbers:
+                    for word in cfg.numbersKeywords[number]:
+                        if word in text:
+                            print("Keyword Found")
+                            time.sleep(25.0)
+                            sem2.acquire()
+                            f = open("output.txt", "r+")
+                            text += f.read()
+                            f.seek(0)
+                            f.truncate()
+                            f.close()
+                            sem2.release()
+                            send_sms.send(text,number,word) #Sends recorded message to list of numbers
+                            text = ""
 
 def main():
 
     global f
+    global numbers
     global listenTrue
     global sem
     global sem2
-    global stopWords
     global threads
+    global text
+    global numberTextToAnalyze
+    global numberOfThreads
+    text = deque()
+    numberTextToAnalyze = 0
     numberOfThreads = 5
-    stopWords = cfg.configuration['keywords']
     listenTrue = True
     sem = threading.Semaphore(value=1)
     sem2 = threading.Semaphore(value=1)
+    numbers = list(cfg.numbersKeywords.keys()) #get all key values
 
     m = sr.Microphone()
 
     f = open('output.txt','w+')
     f.truncate()
     f.close()
+
     object1 = ''
     object2=''
     object3=''
